+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 
 - Add call component that calls a function when started.
 - Add merge component that merges maps when started.
+- Add expand-inline-defs function. This expands component defs that
+  are found inside another component's config. This is convenient for
+  small data adjustments like type conversions and lookups.